Update item-ammo-requirements.json
Added javelin data. Also, this is the entire list of items again (sorry) I had started this before I seen you have tidied up the file by removing all items minus ammo.

Would it be possible to just keep the weapon and ammo items in this file, so that everything is included in order to work out the ammo type and tier in a single file? 

Either that or we split out weapon_type and ammo_tier into the main files? I know